chore(consensus): move tx type and add tests
